Make update-compat.py write Unix-style line-ending
Since we are telling git to not convert line endings on checkout, and
update-compat.py does not touch the .bat files, we can unconditionally
assume that line-endings should be `\n`.

I've factored file-writing out into a separate method so that if we
decide to change this behavior in the future, we only have to change it
in one place.