[CI] Better support for overlaying submodules
The overlay infrastructure now supports declaring that the root
directory of some builds is a subfolder of a given repository.  This
allows overlay-writers to override the repository and branch (and
override the subfolder with "") to indicate that the repository should
be cloned from elsewhere.

This should hopefully provide a relatively seamless experience for
overlay writers who need to update submodules of projects like
fiat-crypto.  In service of that, we split off
ci-{coqutil,kami,riscv_coq,rupicola}